Clean old Go versions
