+++ echo_test.go
@@ -2,6 +2,7 @@ package echo
 
 import (
 	"bytes"
+	stdContext "context"
 	"errors"
 	"net/http"
 	"net/http/httptest"
@@ -460,4 +461,46 @@ func TestHTTPError(t *testing.T) {
 		"code": 12,
 	})
 	assert.Equal(t, "code=400, message=map[code:12]", err.Error())
+}
+
+func TestEchoClose(t *testing.T) {
+	e := New()
+	errCh := make(chan error)
+
+	go func() {
+		errCh <- e.Start(":0")
+	}()
+
+	time.Sleep(200 * time.Millisecond)
+
+	if err := e.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	assert.NoError(t, e.Close())
+
+	err := <-errCh
+	assert.Equal(t, err.Error(), "http: Server closed")
+}
+
+func TestEchoShutdown(t *testing.T) {
+	e := New()
+	errCh := make(chan error)
+
+	go func() {
+		errCh <- e.Start(":0")
+	}()
+
+	time.Sleep(200 * time.Millisecond)
+
+	if err := e.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	ctx, cancel := stdContext.WithTimeout(stdContext.Background(), 10*time.Second)
+	defer cancel()
+	assert.NoError(t, e.Shutdown(ctx))
+
+	err := <-errCh
+	assert.Equal(t, err.Error(), "http: Server closed")
 }
