--- util_go17.go
@@ -1,12 +1,0 @@
-// +build go1.7, !go1.8
-
-package echo
-
-import (
-	"net/url"
-)
-
-// PathUnescape is wraps `url.QueryUnescape`
-func PathUnescape(s string) (string, error) {
-	return url.QueryUnescape(s)
-}
