--- util_go18.go
@@ -1,10 +1,0 @@
-// +build go1.8
-
-package echo
-
-import "net/url"
-
-// PathUnescape is wraps `url.PathUnescape`
-func PathUnescape(s string) (string, error) {
-	return url.PathUnescape(s)
-}
